projects: demo_esp: rework project
Restructure demo_esp8266 project and rename it to demo_esp with the
scope of adding support for multiple platforms / esp versions.
