+++ projects/demo_esp/Makefile
@@ -1,0 +1,9 @@
+# Select the example you want to enable by chossing y for enabling and n for disabling
+BASIC_EXAMPLE = y
+
+# Select the device you want to enable by choosing y for enabling and n for disabling
+include ../../tools/scripts/generic_variables.mk
+
+include src.mk
+
+include ../../tools/scripts/generic.mk
