+++ projects/demo_esp/src/platform/aducm3029/parameters.h
@@ -1,7 +1,7 @@
 /***************************************************************************//**
  *   @file   parameters.h
- *   @brief  Parameters Definitions to configure project.
- *   @author Mihail Chindris (mihail.chindris@analog.com)
+ *   @brief  Definitions specific to aducm3029 platform used by demo_esp
+ *           project.
  *   @author Antoniu Miclaus (antoniu.miclaus@analog.com)
 ********************************************************************************
  * Copyright 2023(c) Analog Devices, Inc.
@@ -37,71 +37,29 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-
 #ifndef __PARAMETERS_H__
 #define __PARAMETERS_H__
 
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-
-#include "aducm3029_irq.h"
+#include "stdio.h"
+#include "platform_init.h"
 #include "aducm3029_timer.h"
+#include "aducm3029_irq.h"
 #include "aducm3029_uart.h"
 
 /******************************************************************************/
 /********************** Macros and Constants Definitions **********************/
 /******************************************************************************/
 
-/* Configure to which mqtt server to connect */
-
-/* Connect to local mqtt broker */
-#define LOCAL			0
-/* Connect to online broker */
-#define TEST_MOSQUITTO		1
-#define MQTT_ECLIPSE		2
-
-#if SERVER_TO_USE==TEST_MOSQUITTO
-#define SERVER_ADDR		"test.mosquitto.org"
-#elif SERVER_TO_USE==MQTT_ECLIPSE
-#define SERVER_ADDR		"mqtt.eclipse.org"
-#else //local
-#define SERVER_ADDR		"10.76.84.237"
-#endif
-
-#define SERVER_PORT		1883
-
-/* Maximum buffer size for a message received from mqtt broker */
-#define BUFF_LEN		200
-
-/* Wifi network parameters */
-#define WIFI_SSID		"WifiSSID"
-#define WIFI_PASS		"Password"
-
-
-/* Default UART parameters for esp8266 */
-#define UART_CONFIG_BAUDRATE	57600
+#define UART_CONFIG_BAUDRATE	115200
 #define UART_CONFIG_IRQ_ID	ADUCM_UART_INT_ID
-
-/* Aducm3029 only have one UART periferal so only 0 can be used */
 #define UART_DEVICE_ID		0
-/* Timer ID to be used for mqtt timeouts */
+#define UART_EXTRA		NULL
+
 #define TIMER_ID		1
-
-/* Mqtt configuration */
-#define MQTT_PUBLISH_TOPIC	"adxl"
-#define MQTT_SUBSCRIBE_TOPIC	"aducm3029_messages"
-#define MQTT_CONFIG_CMD_TIMEOUT	20000
-#define MQTT_CONFIG_VERSION	MQTT_VERSION_3_1
-#define MQTT_CONFIG_KEEP_ALIVE	7200
-#define MQTT_CONFIG_CLIENT_NAME	"aducm3029-client"
-#define MQTT_CONFIG_CLI_USER	NULL
-#define MQTT_CONFIG_CLI_PASS	NULL
-#define MQTT_TIMER_FREQ		1000
-#define MQTT_PAYLOAD_BUFF_LEN	100
-
-/* Time between each scan mq*/
-#define SCAN_SENSOR_TIME	500
+#define TIMER_FREQ		1000
 
 #define IRQ_OPS			&aducm_irq_ops
 #define UART_OPS		&aducm_uart_ops
@@ -110,4 +68,4 @@
 
 extern struct aducm_timer_init_param timer_extra_ip;
 
-#endif // __PARAMETERS_H__
+#endif /* __PARAMETERS_H__ */
