--- projects/demo_esp8266/Makefile
@@ -1,5 +1,0 @@
-include ../../tools/scripts/generic_variables.mk
-
-include src.mk
-
-include ../../tools/scripts/generic.mk
