--- projects/demo_esp8266/builds.json
@@ -1,7 +1,0 @@
-{
-    "aducm3029": {
-        "demo": {
-            "flags" : ""
-        }
-    }
-}
